+++ src/main/java/com/tristankechlo/toolleveling/config/util/ConfigIdentifier.java
@@ -12,22 +12,24 @@ import java.util.function.Function;
 
 public enum ConfigIdentifier implements StringIdentifiable {
 
-    GENERAL("general", "tool_leveling_table.json", ToolLevelingConfig::setToDefault, ToolLevelingConfig::serialize, ToolLevelingConfig::deserialize),
-    ITEM_VALUES("item_values", "item_values.json", ItemValueConfig::setToDefault, ItemValueConfig::serialize, ItemValueConfig::deserialize),
-    COMMANDS("commands", "command_config.json", CommandConfig::setToDefault, CommandConfig::serialize, CommandConfig::deserialize);
+    GENERAL("general", "tool_leveling_table.json", Names.URLs.CONFIG_INFO_GENERAL, ToolLevelingConfig::setToDefault, ToolLevelingConfig::serialize, ToolLevelingConfig::deserialize),
+    ITEM_VALUES("item_values", "item_values.json", Names.URLs.CONFIG_INFO_ITEM_VALUES, ItemValueConfig::setToDefault, ItemValueConfig::serialize, ItemValueConfig::deserialize),
+    COMMANDS("commands", "command_config.json", Names.URLs.CONFIG_INFO_COMMANDS, CommandConfig::setToDefault, CommandConfig::serialize, CommandConfig::deserialize);
 
     @SuppressWarnings("deprecation")
     public static final Codec<ConfigIdentifier> CODEC;
     private final String name;
     private final String fileName;
+    private final String infoUrl;
     private final Runnable resetter;
     private final Function<JsonObject, JsonObject> serializer;
     private final Consumer<JsonObject> deserializer;
 
-    ConfigIdentifier(String name, String fileName, Runnable resetter,
+    ConfigIdentifier(String name, String fileName, String url, Runnable resetter,
                      Function<JsonObject, JsonObject> serializer, Consumer<JsonObject> deserializer) {
         this.name = name;
         this.fileName = fileName;
+        this.infoUrl = url;
         this.resetter = resetter;
         this.serializer = serializer;
         this.deserializer = deserializer;
@@ -56,6 +58,10 @@ public enum ConfigIdentifier implements StringIdentifiable {
 
     public void deserialize(JsonObject json) {
         this.deserializer.accept(json);
+    }
+
+    public String getInfoUrl() {
+        return this.infoUrl;
     }
 
     static {